Fix stoopid bug

+++ src/main/scala/onvifsnapshottaker/photo/FfmpegPhotoMaker.scala
@@ -1,10 +1,13 @@
 package onvifsnapshottaker.photo
 
 import java.nio.file.{Files, Paths}
+import java.util.concurrent.TimeUnit
 
 import onvifsnapshottaker.Config
 
 import scala.util.Try
+import scala.concurrent.duration._
+import scala.language.postfixOps
 
 object FfmpegPhotoMaker extends PhotoMaker {
 
@@ -13,10 +16,14 @@ object FfmpegPhotoMaker extends PhotoMaker {
   private val image = Paths.get(config.getString("image"))
   private val log = Paths.get(config.getString("log"))
 
+  private val timeout = 30 seconds
+
   override def shot(): Try[Array[Byte]] = {
     Try {
       val process = Runtime.getRuntime.exec(s"${script.toString} | tee ${log.toAbsolutePath.toString}")
-      process.waitFor()
+      if(!process.waitFor(timeout.toMillis, TimeUnit.MILLISECONDS)){
+        throw new RuntimeException(s"timeout after ${timeout.toMillis} millis")
+      }
       val bytes = Files.readAllBytes(image)
       Files.delete(image)
       process.destroy()
